+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Global, ThemeProvider } from '@emotion/react';
 import MobileLayout from './components/common/MobileLayout';
-import CartProvider from './context/CartProvider';
+import { CartProvider } from './components/features/cart/context';
 import reset from './global/style/reset';
 import { theme } from './global/style/theme';
 import ShopPage from './pages/shop/ShopPage';
