+++ src/components/features/cart/api/responseMapper.ts
@@ -1,0 +1,13 @@
+import { convertResponseToProduct } from '../../product/api/responseMapper';
+import { Cart } from '../type';
+import { CartItemContent } from './type';
+
+export const convertResponseToCart = ({
+  id,
+  quantity,
+  product,
+}: CartItemContent): Cart => ({
+  id: id.toString(),
+  quantity,
+  product: convertResponseToProduct(product),
+});
