+++ src/components/features/cart/context/index.ts
@@ -1,0 +1,3 @@
+export { default as CartProvider } from './CartProvider';
+export { CartContext } from './CartProvider';
+export { useCartContext } from './useCartContext';
