+++ src/components/features/cart/type.ts
@@ -1,0 +1,7 @@
+import { Product } from '../product';
+
+export interface Cart {
+  id: string;
+  quantity: number;
+  product: Product;
+}
