+++ src/components/features/product/api/index.ts
@@ -1,0 +1,1 @@
+export { getProductList } from './getProductList';
