+++ src/components/features/product/api/responseMapper.ts
@@ -1,5 +1,5 @@
-import { CartItemContent, ProductContent } from './type';
-import { Cart, Product, ProductCategoryType } from '../type';
+import { ProductContent } from './type';
+import { Product, ProductCategoryType } from '../type';
 
 export const convertResponseToProduct = ({
   id,
@@ -15,14 +15,4 @@ export const convertResponseToProduct = ({
   imageUrl: imageUrl ?? 'defaultImage',
   category: (category ?? '전체') as ProductCategoryType,
   quantity,
-});
-
-export const convertResponseToCart = ({
-  id,
-  quantity,
-  product,
-}: CartItemContent): Cart => ({
-  id: id.toString(),
-  quantity,
-  product: convertResponseToProduct(product),
 });
