+++ src/components/features/product/product-card/ProductCard.tsx
@@ -1,11 +1,13 @@
 import { Flex } from '@/components/common';
-import { useCartContext } from '@/context/useCartContext';
+import {
+  addCartItem,
+  deleteCartItem,
+  updateCartItem,
+} from '@/components/features/cart/api';
+import { useCartContext } from '@/components/features/cart/context';
 import { useShopErrorContext } from '@/pages/shop/context/useShopErrorContext';
 import styled from '@emotion/styled';
-import { addCartItem } from '../api/addCartItem';
-import { updateCartItem } from '../api/updateCartItem';
 import AddCartButton from './AddCartButton';
-import { deleteCartItem } from '../api/deleteCartItem';
 
 interface ProductProps {
   id: string;
