+++ src/components/features/product/type.ts
@@ -1,9 +1,3 @@
-export interface Cart {
-  id: string;
-  quantity: number;
-  product: Product;
-}
-
 export interface Product {
   id: string;
   name: string;
